refactor: Finish code cleanup & optimization

+++ Engine/Code/Source/Rendering/Vulkan/VulkanSwapChain.cpp
@@ -81,7 +81,7 @@ QueueFamilyIndices VulkanSwapChain::GetQueueFamilies(const VkPhysicalDevice a_de
     for (unsigned int i = 0; i < l_queueFamilyList.size(); ++i)
     {
         if (l_queueFamilyList[i].queueCount > 0 && l_queueFamilyList[i].queueFlags & VK_QUEUE_GRAPHICS_BIT)
-            l_indices.graphicsFamily = i;
+            l_indices.graphicsFamily = static_cast<int>(i);
 
         VkBool32 l_presentationSupport{ false };
         vkGetPhysicalDeviceSurfaceSupportKHR(a_device, 1, a_surface, &l_presentationSupport);
